Fix issues with erlbin serialization.
All messages with arguments that did not include both of them failed
in awre. Those cases have not been implemented as wamper converted
them to always include those Arguments, yet not with the new erlbin
serialization.
An example would be {error,call,#{},no_such_procedure}.
Fix #4

Another small bug was that the wrong reference was searched when
returning from a connection.
Fix #3

Thanks  @wcreek for filing and already fixing #3.